+++ reed-solomon-novelpoly/src/novel_poly_basis/encode.rs
@@ -1,0 +1,56 @@
+use super::*;
+
+use itertools::Itertools;
+
+/// Bytes shall only contain payload data
+pub fn encode_sub(bytes: &[u8], n: usize, k: usize) -> Result<Vec<Additive>> {
+	assert!(is_power_of_2(n), "Algorithm only works for 2^i sizes for N");
+	assert!(is_power_of_2(k), "Algorithm only works for 2^i sizes for K");
+	assert!(bytes.len() <= k << 1);
+	assert!(k <= n / 2);
+
+	// must be power of 2
+	let dl = bytes.len();
+	let l = if is_power_of_2(dl) {
+		dl
+	} else {
+		let l = log2(dl);
+		let l = 1 << l;
+		let l = if l >= dl { l } else { l << 1 };
+		l
+	};
+	assert!(is_power_of_2(l));
+	assert!(l >= dl);
+
+	// pad the incoming bytes with trailing 0s
+	// so we get a buffer of size `N` in `GF` symbols
+	let zero_bytes_to_add = n * 2 - dl;
+	let data: Vec<Additive> = bytes
+		.into_iter()
+		.copied()
+		.chain(std::iter::repeat(0u8).take(zero_bytes_to_add))
+		.tuple_windows()
+		.step_by(2)
+		.map(|(a, b)| Additive(u16::from_be_bytes([a, b])))
+		.collect::<Vec<Additive>>();
+
+	// update new data bytes with zero padded bytes
+	// `l` is now `GF(2^16)` symbols
+	let l = data.len();
+	assert_eq!(l, n);
+
+	let mut codeword = data.clone();
+	assert_eq!(codeword.len(), n);
+
+	encode_low(&data[..], k, &mut codeword[..], n);
+
+	Ok(codeword)
+}
+
+
+pub fn encode(bytes: &[u8], validator_count: usize) -> Result<Vec<WrappedShard>> {
+	let params = CodeParams::derive_from_validator_count(validator_count)?;
+
+	let rs = params.make_encoder();
+	rs.encode(bytes)
+}
